refactor: turn mockFetch into a full module mock
This should make it harder to make a mistake mocking `cross-fetch` again
