+++ test/mocks/mock-fetch.js
@@ -1,6 +1,7 @@
 const TextEncoder = require('util').TextEncoder;
-const {Headers} = require('cross-fetch');
+const crossFetch = require('cross-fetch');
 
+const Headers = crossFetch.Headers;
 const successText = 'successful response';
 
 /**
@@ -58,7 +59,11 @@ const mockFetch = (resource, options) => {
     return Promise.resolve(results);
 };
 
+// Mimic the cross-fetch module, but replace its `fetch` with `mockFetch` and add a few extras
 module.exports = {
+    ...crossFetch, // Headers, Request, Response, etc.
+    default: mockFetch,
+    fetch: mockFetch,
     mockFetch,
     successText
 };
