+++ test/unit/fetch-tool.js
@@ -1,19 +1,14 @@
 const tap = require('tap');
 const TextDecoder = require('util').TextDecoder;
-const crossFetch = require('cross-fetch');
 
-const {mockFetch, successText} = require('../mocks/mockFetch.js');
+const mockFetch = require('../mocks/mock-fetch.js');
 
 /**
  * This is the real FetchTool, but the 'cross-fetch' module has been replaced with the mockFetch function.
  * @type {typeof import('../../src/FetchTool')}
  */
 const FetchTool = tap.mock('../../src/FetchTool', {
-    'cross-fetch': {
-        ...crossFetch, // Headers, etc.
-        default: mockFetch,
-        fetch: mockFetch
-    }
+    'cross-fetch': mockFetch
 });
 
 tap.test('send success returns response.text()', t => {
@@ -21,7 +16,7 @@ tap.test('send success returns response.text()', t => {
 
     return t.resolves(
         tool.send({url: '200'}).then(result => {
-            t.equal(result, successText);
+            t.equal(result, mockFetch.successText);
         })
     );
 });
@@ -40,7 +35,7 @@ tap.test('get success returns Uint8Array.body(response.arrayBuffer())', t => {
 
     return t.resolves(
         tool.get({url: '200'}).then(result => {
-            t.equal(decoder.decode(result), successText);
+            t.equal(decoder.decode(result), mockFetch.successText);
         })
     );
 });
